requirements-audio.txt
# Core backend dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
requests==2.32.3

# Document processing
PyPDF2==3.0.1
python-docx==1.1.2

# Frontend dependencies
streamlit==1.39.0
pandas==2.2.3
plotly==5.24.1

# Audio transcription (NEW)
openai-whisper==20231117
pydub==0.25.1

# Optional: OpenAI API for cloud transcription
# Uncomment if you want to use OpenAI Whisper API instead of local
# openai==1.3.0

# Optional: Advanced speaker diarization
# Uncomment for better speaker detection
# pyannote.audio==3.1.1

# Note: FFmpeg must be installed separately
# macOS: brew install ffmpeg
# Linux: apt install ffmpeg
# Windows: Download from ffmpeg.org

# Note: Ollama must be installed separately
# Visit https://ollama.ai for installation instructions

